Time: 8 ms (60.86%) | Memory: 8.5 MB (76.95%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         if(x<0) return false;
-        long long rev = 0;
-        long long tem = x;
+        long rev = 0;
+        long tem = x;
         while(tem != 0){
             int remainder = tem % 10;
             rev = rev * 10 + remainder;
